app/requirements.txt
Django>=4.2,<5
gunicorn
psycopg2-binary
django-import-export
openpyxl>=3.0.0
pandas>=1.3.0